added local and remote modes

+++ skel_script.sh
@@ -1,24 +1,36 @@
 #!/bin/bash
 
-GIT_URL="$1"
+MODE="$1"
+GIT_URL="$2"
 HASH_LIST=$(mktemp)
 DIFF_FILE=$(mktemp)
 KEY_DIR=$(mktemp -d)
 
-trufflehog --regex  --entropy=False $GIT_URL | grep "Reason: .* key" -A 2 | grep "Hash:" | awk -F" " '{print $2}' | sort -u | sed -r "s/\x1B\[([0-9]{1,3}(;[0-9]{1,2})?)?[mGK]//g" > $HASH_LIST
+if [[ "$MODE" == "local" ]]; then
+	trufflehog git_url --repo_path $GIT_URL --regex --entropy=False | grep "Reason: .* key" -A 2 | grep "Hash:" | awk -F" " '{print $2}' | sort -u | sed -r "s/\x1B\[([0    -9]{1,3}(;[0-9]{1,2})?)?[mGK]//g" > $HASH_LIST	
+else
+	trufflehog --regex  --entropy=False $GIT_URL | grep "Reason: .* key" -A 2 | grep "Hash:" | awk -F" " '{print $2}' | sort -u | sed -r "s/\x1B\[([0-9]{1,3}(;[0-9]{1,2})?)?[mGK]//g" > $HASH_LIST
+fi
+
 
 if [[ "$(wc -l $HASH_LIST | awk -F" " '{print $1}')" -ne "0" ]]; then
-	GIT_DIR=$(mktemp -d)
-	git clone $GIT_URL $GIT_DIR
-	cd $GIT_DIR
+	if [[ "$MODE" == "remote" ]]; then
+		GIT_DIR=$(mktemp -d)
+		git clone $GIT_URL $GIT_DIR
+		cd $GIT_DIR
+	else
+		cd $GIT_URL
+	fi
 	cat $HASH_LIST | while IFS= read -r line; do 		
 		DIFF_FILE=$(mktemp)
-		git show $line | sed -r "s/^([^-+ ]*)[-+ ]/\\1/" > $DIFF_FILE; 
+		git show $line | sed -r "s/^([^-+ ]*)[-+ ]/\\1/" | sed -r "s/\x1B\[([0-9]{1,3}(;[0-9]{1,2})?)?[mGK]//g" > $DIFF_FILE; 
 		awk '/^-----BEGIN [A-Z]* PRIVATE (KEY|KEY BLOCK)-----$/{flag=1}/^-----END [A-Z]* PRIVATE (KEY|KEY BLOCK)-----$/{print;flag=0}flag' $DIFF_FILE > ${KEY_DIR}/private_key_$(echo $GIT_URL | awk -F/ '{print $NF}' | awk -F. '{print $1}')_${line}
 		rm -f $DIFF_FILE
 	done
 	cd /tmp/
-	rm -rf $GIT_DIR
+	if [[ "$MODE" == "remote" ]]; then
+		rm -rf $GIT_DIR
+	fi
 	echo "Collected keys at: $KEY_DIR"
 fi
